requirements.txt
streamlit
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
chromadb
openai
torch
llama-index
spacy
scikit-learn
imapclient
google-auth-oauthlib
pydantic
transformers==4.39.3
python-jose[cryptography]
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
beautifulsoup4>=4.9.3
werkzeug>=2.0.0  
chromadb
pydantic
sentence-transformers